seperate world rendering from react component
