+++ client/components/index.js
@@ -1,5 +1,4 @@
 
 export {default as UserHome} from './user-home'
 export {default as World} from './World'
-export {default as World2} from './World2'
 export {Login, Signup} from './auth-form'
